Fix keyboard shortcuts
There were some keyboard shortcuts in the GIMP 2.0 version of the
repository that were not properly transferred when the repository was
upgraded to support GIMP 3.0. Also, there were some keyboard shortcuts
that were too narrowly defined or simply were not working, so this
commit fixes all of these issues.

Fixes #23, #77